Update 2025-07-28-ganaanaam-tva-ganapati.md
In correspondence to the changed recording, I had to edit the javascript that highlights lines as they are being chanted. This should be the last of content changes for this post and the succeeding posts will resemble this but, of course, with different chants. If new enhancements or ideas need to added, further commits to this page will be made.